Implement search box on index webpage
Search box works for finding exact news item by the text in a title or description
field. Controller modified in that way: if there is a search parameter in http request
it will switch the case and take into consideration search parameter.
If there is no search paremeter, the index action will display all news.