Remove check statement for pictures
Loading of the index page was slow because there is checking state for
displaying image

If there is no picture added to the news item, index page return "nil".
So if stament checks when the item have a picture or not.
On opening index page on loading each item was send a request, that slows down
whole loading of the page.

To reduce the amount of requests, the check statement was changed on
catch block