Add solution for heroku display images
Heroku cloud service after some period of time reload server and after
that all pictures links in web page, that was added through the Active
Storage in PostgreSQL database disapear.

As a solution I use the local folder for storaging images. Also I
leave the  possibility to add pictures through the Active Storage
to show that it works for other servers. This happens due to heroku own
server properties.